Extracted keyboard controls, fixed bug in OscDimmedChannel, fixed bug in Ticker/Tickable

+++ src/main/kotlin/de/randombyte/lighthub/show/flows/intensity/IntensityFlow.kt
@@ -16,7 +16,7 @@ object IntensityFlow : Flow<MasterDimmerFeature>(acceptedDevices = ledBars) {
 
     override fun onTick(tick: ULong, device: MasterDimmerFeature) {
         if (mode == Sawtooth) {
-            device.masterDimmer = DMX_RANGE.last * getPercentUntilNextChange<IntensityAutoPatternsConfig>(tick, device as Device)
+            device.masterDimmer = (DMX_RANGE.last * getPercentUntilNextChange<IntensityAutoPatternsConfig>(tick, device as Device)).toInt()
         }
     }
 }
